- user customer JPA Mapping - Refactor Bootstrap Data
